+openCV
Added opencv: tracking of the bobber + creating a rectangular around it. The programm then checks for this rect. to change, and when it changes, aka bobber moves, the program changes cursors position, clicks on the bobber and starts fishing again.
Still need a bit of work: if there's no baiting, the prog. wont click and so it wont start a new fishing sequance. Need to add some external timer, that will check if programm did something in last 20 secs, and if not => start fishing again.
Or check for the rectangular. If there's no rect. for ~2 secs than start again